refactor: 테스트 코드를 domains 폴더로 분리
